+++ .gitignore
@@ -131,3 +131,7 @@ dmypy.json
 # Texteditors
 .idea/
 .vscode/
+
+# Temp
+temptest.py
+gt.py
