+++ requirements_old.txt
@@ -1,23 +1,15 @@
 autopep8==1.7.0
 certifi==2022.9.24
-cffi==1.15.1
 charset-normalizer==2.1.1
 click==8.1.3
-colorama==0.4.5
-Deprecated==1.2.13
 Flask==2.2.2
-gunicorn==20.1.0
 idna==3.4
 itsdangerous==2.1.2
 Jinja2==3.1.2
 MarkupSafe==2.1.1
 pycodestyle==2.9.1
-pycparser==2.21
-PyGithub==1.55
-PyJWT==2.5.0
-PyNaCl==1.5.0
 requests==2.28.1
 toml==0.10.2
 urllib3==1.26.12
 Werkzeug==2.2.2
-wrapt==1.14.1
+gunicorn==20.1.0
